DEV: Scroll to top of page on livestream topic loaded (#15)
Chat scrolls to the bottom of the page when the chat channel is loaded. This is required for the chat message positioning to be correct. We need to scroll to the top of the page after the chat channel is loaded to avoid the page being rendered and the viewport being scrolled to the bottom.
      
This is not an ideal solution, but it's the best we can do for now

# Desktop
https://github.com/discourse/discourse-livestream/assets/50783505/44250754-64e3-44b5-9976-6131f6779a4e

# Mobile
https://github.com/discourse/discourse-livestream/assets/50783505/5f22674c-0854-4adf-a5e3-0c45f69f8196